chore: Update npm dependencies
This commit updates the npm dependencies in the project. It includes updating the "cors" package to version 2.8.5 and adding the "@types/cors" package with version 2.8.17 as a dev dependency. The updated dependencies will improve the functionality and compatibility of the application.

This commit message follows the established convention of using a prefix to indicate the type of change, followed by a concise and descriptive message.